improve integer support
