+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## [_Unreleased_](https://github.com/freckle/debug-print/compare/v0.0.0.0...main)
+## [_Unreleased_](https://github.com/freckle/debug-print/compare/v0.1.0.0...main)
+
+## [v0.1.0.0](https://github.com/freckle/nonempty-zipper/compare/v0.0.0.0...v0.1.0.0)
+
+The parameter of `DebugPrintValueInt` is changed from `Int` to `Integer`.
+
+Added `ToDebugPrintValue` instances for various numeric types.
+
+Adds the `DebugInteger` newtype for deriving-via.
 
 ## [v0.0.0.0](https://github.com/freckle/debug-print/tree/v0.0.0.0)
