+++ library/DebugPrint/Integer.hs
@@ -1,0 +1,5 @@
+module DebugPrint.Integer
+  ( DebugInteger (..)
+  ) where
+
+import DebugPrint.Core
