+++ package.yaml
@@ -1,5 +1,5 @@
 name: debug-print
-version: 0.0.0.0
+version: 0.1.0.0
 maintainer: Freckle Education
 category: Debug
 github: freckle/debug-print
@@ -45,8 +45,17 @@ library:
   dependencies:
     - aeson
     - containers
+    - internal
     - text
-    - vector
+
+internal-libraries:
+  internal:
+    source-dirs: internal
+    dependencies:
+      - containers
+      - text
+      - vector
+
 tests:
   readme:
     main: README.lhs
